refactor: detailed-schedule 폴더 위치 이동
