+++ src/features/detailed-schedule/__stories__/detailed-schedule-color-badge.stories.tsx
@@ -1,4 +1,4 @@
-import DetailedScheduleColorBadge from './detailed-schedule-color-badge'
+import { DetailedScheduleColorBadge } from '../ui'
 
 import type { Meta, StoryObj } from '@storybook/react-vite'
