+++ src/features/detailed-schedule/ui/detailed-schedule-color-badge.tsx
