+++ src/features/detailed-schedule/ui/detailed-schedule-date.tsx
@@ -1,5 +1,4 @@
-import Text from '../text/text'
-
+import { Text } from '@/shared/ui'
 interface Props {
   date: string
   weekday: string
