+++ src/features/detailed-schedule/ui/detailed-schedule-header.tsx
