+++ src/features/detailed-schedule/ui/detailed-schedule-list-card.tsx
