+++ src/features/detailed-schedule/ui/index.ts
