+++ src/features/share-schedule/ui/share-detailed-schedule-list.tsx
@@ -4,9 +4,9 @@ import { useNavigate, useParams } from 'react-router'
 import { useShareScheduleByUserId } from '@/entities/schedule/hooks'
 import { formatScheduleTime } from '@/entities/utils/format-schedule-time'
 import { PrivateScheduleModal } from '@/features/calendar/ui'
+import { DetailedScheduleListCard } from '@/features/detailed-schedule/ui'
 import { IconInvite } from '@/shared/assets/icons'
 import { FloatButton } from '@/shared/ui'
-import { DetailedScheduleListCard } from '@/shared/ui/detailed-schedule'
 
 export default function ShareDetailedScheduleList() {
   const navigate = useNavigate()
