+++ src/features/share-schedule/ui/share-detailed-schedule.tsx
@@ -1,8 +1,8 @@
 import { useParams } from 'react-router'
 
 import { useDetailedSchedule } from '@/entities/schedule/hooks'
-import { getRepeatText } from '@/entities/utils/format-repeat-text'
-import DetailedScheduleCard from '@/shared/ui/detailed-schedule/detailed-schedule-card'
+import { getRepeatText } from '@/entities/utils'
+import { DetailedScheduleCard } from '@/features/detailed-schedule/ui'
 
 export default function ShareDetailedSchedule() {
   const { scheduleId } = useParams()
